+++ src/InternalVarDumper.php
@@ -4,12 +4,25 @@ namespace Awesomite\VarDumper;
 
 class InternalVarDumper implements VarDumperInterface
 {
+    protected $displayPlaceInCode;
+    
+    private $shift;
+    
+    public function __construct($displayPlaceInCode = false, $stepShift = 0)
+    {
+        $this->displayPlaceInCode = $displayPlaceInCode;
+        $this->shift = $stepShift;
+    }
+
     public function dump($var)
     {
         $iniKey = 'xdebug.overload_var_dump';
         $previousVal = ini_get($iniKey);
         ini_set($iniKey, 0);
 
+        if ($this->displayPlaceInCode) {
+            $this->dumpPlaceInCode(0);
+        }
         var_dump($var);
 
         ini_set($iniKey, $previousVal);
@@ -23,5 +36,24 @@ class InternalVarDumper implements VarDumperInterface
         ob_end_clean();
 
         return $result;
+    }
+    
+    protected function dumpPlaceInCode($number)
+    {
+        $options = version_compare(PHP_VERSION, '5.3.6') >= DEBUG_BACKTRACE_IGNORE_ARGS ? 0 : false;
+        $stackTrace = debug_backtrace($options);
+        $num = 1 + $number + $this->shift;
+
+        // @codeCoverageIgnoreStart
+        if (!isset($stackTrace[$num])) {
+            return;
+        }
+        // @codeCoverageIgnoreEnd
+
+        $step = $stackTrace[$num];
+
+        if (isset($step['file']) && $step['file']) {
+            echo $step['file'] . (isset($step['line']) ? ':' . $step['line'] : '') . ":\n";
+        }
     }
 }