+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0
+
+* Added `\Awesomite\VarDumper\Dumper`
+* Added arguments `$displayPlaceInCode` and `$stepShift` to
+`\Awesomite\VarDumper\InternalVarDumper::__construct`
+and `\Awesomite\VarDumper\LightVarDumper::__construct`
+
 ## 0.1.0
 
 * Initial public release